Switch app backend from Google Sheets to Airtable
It types the data properties a bit better. Removes the intermediate query for available sheet ids. But mostly the switch is for a friendlier admin CMS.

Yes, that's an API key in the source. If Ember's going to make API requests, it needs credentials at compile time. The key's permissions are read-only, so it's okay for now. To really lock it down to the specifics of this app, later, write a proxy that gets the API key from the server side environment.