Paginate Airtable results until you fetch them all
They have a cap of 100 rows and we have slightly more than that. We'd like to filter & sort them all in memory for now. But they have API filtering and sorting when we need it.

Exercise this code path with Mirage by splitting the database into a page of 8 records and a page of 2.