Restore direct link scrolling
this.controllerFor(...) perhaps was looking up the wrong controller to feed the card-list component. Anyway, things are a lot easier if you set the property on the model directly. To make sure you have an Ember.Object to call .set on, fully build up the model in the, well, model hook.